Merge the current default history while selecting its newer build and workflow paths over the 191-commit-old draft. The preserved checkpoint remains in history, but no obsolete parallel pipeline survives in the resulting tree; museums need labels, not production entry points.

合入現時預設歷史，揀返新版 build 同 workflow 路徑，唔畀落後 191 個 commits 嘅草稿搶軚。保留點仍然完整喺 history，但結果唔會留低過期平行 pipeline；博物館要標籤，唔係 production 入口。

+++ src/VBox/Runtime/r0drv/nt/initterm-r0drv-nt.cpp
@@ -497,7 +497,7 @@ DECLHIDDEN(int) rtR0InitNative(void)
             /* Try all sets: iBest -> End; iBest -> Start. */
             bool    fDone = false;
             int32_t i     = iBest;
-            while (   i < RT_ELEMENTS(g_artNtSdbSets)
+            while (   i < (int32_t)RT_ELEMENTS(g_artNtSdbSets)
                    && !(fDone = rtR0NtTryMatchSymSet(&g_artNtSdbSets[i], pbPrcb, u.szVendor, &OsVerInfo)))
                 i++;
             if (!fDone)
